requirements.txt
Flask==0.9
SQLAlchemy==0.7.9
facebook-sdk==0.4.0
requests==1.1.0
redis==2.7.2
